CORE-5001 Extract TestSerializationService to a dedicated class
