+++ libs/ledger/ledger-consensual-impl/src/test/kotlin/net/corda/ledger/consensual/impl/helper/TestSerializationService.kt
@@ -1,0 +1,64 @@
+package net.corda.ledger.consensual.impl.helper
+
+import net.corda.internal.serialization.SerializationContextImpl
+import net.corda.internal.serialization.SerializationServiceImpl
+import net.corda.internal.serialization.amqp.DefaultDescriptorBasedSerializerRegistry
+import net.corda.internal.serialization.amqp.DescriptorBasedSerializerRegistry
+import net.corda.internal.serialization.amqp.DeserializationInput
+import net.corda.internal.serialization.amqp.SerializationOutput
+import net.corda.internal.serialization.amqp.SerializerFactory
+import net.corda.internal.serialization.amqp.SerializerFactoryBuilder
+import net.corda.internal.serialization.amqp.amqpMagic
+import net.corda.internal.serialization.amqp.currentSandboxGroup
+import net.corda.libs.packaging.core.CpkMetadata
+import net.corda.sandbox.SandboxGroup
+import net.corda.serialization.SerializationContext
+import net.corda.v5.application.serialization.SerializationService
+import org.osgi.framework.Bundle
+
+//TODO(Deduplicate with net.corda.internal.serialization.amqp.testutils)
+
+private class MockSandboxGroup(private val classLoader: ClassLoader = ClassLoader.getSystemClassLoader()) :
+    SandboxGroup {
+    override val metadata: Map<Bundle, CpkMetadata> = emptyMap()
+
+    override fun loadClassFromMainBundles(className: String): Class<*> =
+        Class.forName(className, false, classLoader)
+    override fun <T : Any> loadClassFromMainBundles(className: String, type: Class<T>): Class<out T> =
+        Class.forName(className, false, classLoader).asSubclass(type)
+    override fun getClass(className: String, serialisedClassTag: String): Class<*> = Class.forName(className)
+    override fun getStaticTag(klass: Class<*>): String = "S;bundle;sandbox"
+    override fun getEvolvableTag(klass: Class<*>) = "E;bundle;sandbox"
+}
+
+private val testSerializationContext = SerializationContextImpl(
+    preferredSerializationVersion = amqpMagic,
+    properties = mutableMapOf(),
+    objectReferencesEnabled = false,
+    useCase = SerializationContext.UseCase.Testing,
+    encoding = null,
+    sandboxGroup = MockSandboxGroup()
+)
+
+private fun testDefaultFactoryNoEvolution(
+    descriptorBasedSerializerRegistry: DescriptorBasedSerializerRegistry =
+        DefaultDescriptorBasedSerializerRegistry()
+): SerializerFactory =
+    SerializerFactoryBuilder.build(
+        testSerializationContext.currentSandboxGroup(),
+        descriptorBasedSerializerRegistry = descriptorBasedSerializerRegistry,
+        allowEvolution = false
+    )
+
+class TestSerializationService {
+    companion object{
+        fun getTestSerializationService() : SerializationService {
+            val factory = testDefaultFactoryNoEvolution()
+            val output = SerializationOutput(factory)
+            val input = DeserializationInput(factory)
+            val context = testSerializationContext
+
+            return SerializationServiceImpl(output, input, context)
+        }
+    }
+}
